+++ app/src/main/AndroidManifest.xml
@@ -21,6 +21,9 @@
 
         <activity
             android:name=".MainActivity"
+
+            android:theme="@style/Theme.AppCompat.NoActionBar"
+
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
