+++ app/src/main/java/com/example/wlmap/Graph.kt
@@ -3,12 +3,8 @@ package com.example.wlmap
 import android.content.ContentValues
 import android.util.Log
 import com.mapbox.geojson.Point
-import com.mapbox.maps.logD
-import com.mapbox.maps.logE
-import org.w3c.dom.Node
 import java.util.PriorityQueue
 import kotlin.math.*
-import kotlin.random.Random
 
 
 class Graph {
