+++ app/src/main/java/com/example/wlmap/UpdatableMap.kt
@@ -1,0 +1,6 @@
+package com.example.wlmap
+
+interface UpdatableMap {
+    fun updateLocation(lat: Double, lon: Double)
+
+}
