+++ app/src/main/res/drawable/color_sidebar.xml
@@ -1,0 +1,9 @@
+<shape xmlns:android="http://schemas.android.com/apk/res/android"
+    android:shape="rectangle">
+    <gradient
+        android:angle="135"
+        android:endColor="#0468FF"
+        android:centerColor="#2EC8FB"
+        android:startColor="#A8E0FB"
+        android:type="linear" />
+</shape>
