+++ app/src/main/res/drawable/ic_sidebar_home.xml
@@ -1,0 +1,8 @@
+<!--Creating a home icon for the Home button on the sidebar -->
+<vector xmlns:android="http://schemas.android.com/apk/res/android" android:height="24dp" android:tint="#0051E2" android:viewportHeight="24" android:viewportWidth="24" android:width="24dp">
+      
+    <path android:fillColor="@android:color/white" android:pathData="M1,11l0,10l5,0l0,-6l4,0l0,6l5,0l0,-10l-7,-5z"/>
+      
+    <path android:fillColor="@android:color/white" android:pathData="M10,3v1.97l7,5V11h2v2h-2v2h2v2h-2v4h6V3H10zM19,9h-2V7h2V9z"/>
+    
+</vector>
