+++ app/src/main/res/drawable/rounded_corner_box.xml
@@ -1,0 +1,6 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape xmlns:android="http://schemas.android.com/apk/res/android"
+    android:shape="rectangle">
+    <corners android:radius="24dp"/>
+    <stroke android:width="2dip" android:color="@color/black"/>
+</shape>
