+++ app/src/main/res/layout/fragment_home.xml
@@ -1,0 +1,50 @@
+<?xml version="1.0" encoding="utf-8"?>
+<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    android:gravity="center"
+    android:padding="30dp"
+    tools:context="com.example.wlmap.HomeFragment">
+
+    <!--    We align the fill-in features at the center of the screen-->
+    <LinearLayout
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:gravity="center"
+        android:orientation="vertical">
+
+        <TextView
+            android:id="@+id/send_data_title"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_marginBottom="60dp"
+            android:text="Welcome to WLMap!\n\n\nChoose the following options to continue"
+            android:textAlignment="center"
+
+            android:textSize="30sp" />
+
+        <!-- Navigation Button -->
+        <Button
+            android:id="@+id/nav_button"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:layout_marginTop="32dp"
+            android:padding="10dp"
+            android:text="Navigation"
+            android:textSize="16sp" />
+
+        <Button
+            android:id="@+id/data_collection_button"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:layout_marginTop="32dp"
+            android:padding="10dp"
+            android:text="Data Collection"
+            android:textSize="16sp" />
+
+
+    </LinearLayout>
+
+</RelativeLayout>
