+++ app/src/main/res/values/strings.xml
@@ -1,3 +1,23 @@
 <resources>
+    <!--    Creating names/labels for the main screen-->
     <string name="app_name">WLMap</string>
+    <string name="open_home">Open Navigation Home</string>
+    <string name="close_home">Close Navigation Home</string>
+
+    <!--    Creating the label for the sidebar -->
+    <string name="navigation_drawer_open">Open navigation drawer</string>
+    <string name="navigation_drawer_close">Close navigation drawer</string>
+    <string name="nav_header_title">UBIT Name</string>
+    <string name="nav_header_subtitle">ubit@buffalo.edu</string>
+    <string name="nav_header_desc">Navigation header</string>
+    <string name="action_settings">Settings</string>
+
+    <!--    Creating the name for the buttons in the sidebar -->
+    <string name="name_sidebar_home">Home</string>
+    <string name="name_sidebar_map">Navigation</string>
+    <string name="name_sidebar_data">Share Data</string>
+    <string name="name_sidebar_setting">Setting</string>
+    <string name="name_sidebar_logout">Log Out</string>
+    <!-- TODO: Remove or change this placeholder text -->
+    <string name="hello_blank_fragment">Hello blank fragment</string>
 </resources>
